+++ src/ts/Views/MRAID.ts
@@ -20,7 +20,7 @@ export interface IOrientationProperties {
 
 export class MRAID extends View {
 
-    public readonly onClick = new Observable0();
+    public readonly onClick = new Observable1<string>();
     public readonly onReward = new Observable0();
     public readonly onSkip = new Observable0();
     public readonly onClose = new Observable0();
@@ -158,6 +158,15 @@ export class MRAID extends View {
         });
     }
 
+    public setViewableState(viewable: boolean) {
+        if(this._iframeLoaded) {
+            this._iframe.contentWindow.postMessage({
+                type: 'viewable',
+                value: viewable
+            }, '*');
+        }
+    }
+
     private onIframeLoaded() {
         this._iframeLoaded = true;
 
@@ -273,15 +282,7 @@ export class MRAID extends View {
     private onMessage(event: MessageEvent) {
         switch(event.data.type) {
             case 'open':
-                this.onClick.trigger();
-                if(this._nativeBridge.getPlatform() === Platform.IOS) {
-                    this._nativeBridge.UrlScheme.open(encodeURI(event.data.url));
-                } else if(this._nativeBridge.getPlatform() === Platform.ANDROID) {
-                    this._nativeBridge.Intent.launch({
-                        'action': 'android.intent.action.VIEW',
-                        'uri': encodeURI(event.data.url) // todo: these come from 3rd party sources, should be validated before general MRAID support
-                    });
-                }
+                this.onClick.trigger(encodeURI(event.data.url));
                 break;
 
             case 'close':